integration-tests/requirements.txt
# Integration Test Dependencies

# Testing framework
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # parallel test execution

# HTTP client
httpx==0.26.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data validation
pydantic==2.6.0

# Environment
python-dotenv==1.0.1

# Logging and debugging
structlog==24.1.0

# Utilities
faker==22.6.0  # generate test data
